// packages/react/src/theme/palette.generated.ts
export const shades = {
  white: '#ffffff',
  whiteTransparent: {
    '0': '#ffffff00',
    '5': '#ffffff0d',
    '10': '#ffffff1a',
    '15': '#ffffff26',
    '20': '#ffffff33',
    '25': '#ffffff40',
    '30': '#ffffff4d',
    '35': '#ffffff59',
    '40': '#ffffff66',
    '45': '#ffffff73',
    '50': '#ffffff80',
    '55': '#ffffff8c',
    '60': '#ffffff99',
    '65': '#ffffffa6',
    '70': '#ffffffb3',
    '75': '#ffffffbf',
    '80': '#ffffffcc',
    '85': '#ffffffd9',
    '90': '#ffffffe6',
    '95': '#fffffff2',
  },
  black: '#000000',
  blackTransparent: {
    '0': '#00000000',
    '5': '#0000000d',
    '10': '#0000001a',
    '15': '#00000026',
    '20': '#00000033',
    '25': '#00000040',
    '30': '#0000004d',
    '35': '#00000059',
    '40': '#00000066',
    '45': '#00000073',
    '50': '#00000080',
    '55': '#0000008c',
    '60': '#00000099',
    '65': '#000000a6',
    '70': '#000000b3',
    '75': '#000000bf',
    '80': '#000000cc',
    '85': '#000000d9',
    '90': '#000000e6',
    '95': '#000000f2',
  },
} as const;